requirements.txt
django
djangorestframework
django-cors-headers
torch
torchaudio
numpy
streamlit
requests
python-dotenv
soundfile
transformers
chatterbox-tts